backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg[binary]>=3.1.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
redis>=5.0.0
httpx>=0.26.0
openai>=1.12.0
python-dotenv>=1.0.0
slowapi>=0.1.9
